# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rimsys OpenApi",
        "description": "RimSys OpenApi description",
        "contact": {
            "email": "tunjiebifemi@gmail.com.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login (Authentitication)",
                "operationId": "login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Registration",
                "operationId": "register",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter your name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter your Email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Enter your password",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Enter your password confirmation",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout user and invalidate token",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get User's Documents",
                "description": "Fetches documents by currently logged in user",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a Documents",
                "operationId": "store",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "document_name": {
                                        "description": "Enter name of the document",
                                        "type": "string"
                                    },
                                    "document_type": {
                                        "description": "Enter document type",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/documents/{document}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Find document by ID",
                "description": "Returns a single document",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID of document to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Deletes a document",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "document",
                        "in": "path",
                        "description": "Document id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Documents"
                ],
                "summary": "Updates a document",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID of document that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "document_name": {
                                        "description": "Updated name of the document",
                                        "type": "string"
                                    },
                                    "document_type": {
                                        "description": "Updated type of the document",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum Security Laravel",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}